Hide/show poll view (#43)
* Hide functionality added

By default, results are hidden. Unhide and hide with the hide/unhide button. State of hide/unhide does not change when we create a new question or when we go to next question. The user has to manually hide or unhide as they desire.

* lib/constants addition & wording toggle

Previously a const variable was defined within this file but it was moved to lib/constants

instead of "unhide" button now says "show"

* centered showResults button

The show/hide button has been centered.

* remove whitespace padding

Remove the padding of the cardcontent container so that there is no white space when hiding poll results